+++ zebra/zebra_nhg.h
@@ -41,7 +41,6 @@ struct nh_grp {
 
 PREDECL_RBTREE_UNIQ(nhg_connected_tree);
 
-
 /*
  * Hashtables contiaining entries found in `zebra_router`.
  */
@@ -111,21 +110,6 @@ struct nhg_hash_entry {
 /* Was this one we created, either this session or previously? */
 #define ZEBRA_NHG_CREATED(NHE) ((NHE->type) == ZEBRA_ROUTE_NHG)
 
-/* Abstraction for connected trees */
-struct nhg_connected {
-	struct nhg_connected_tree_item tree_item;
-	struct nhg_hash_entry *nhe;
-};
-
-static int nhg_connected_cmp(const struct nhg_connected *con1,
-			     const struct nhg_connected *con2)
-{
-	return (con1->nhe->id - con2->nhe->id);
-}
-
-DECLARE_RBTREE_UNIQ(nhg_connected_tree, struct nhg_connected, tree_item,
-		    nhg_connected_cmp);
-
 
 enum nhg_ctx_op_e {
 	NHG_CTX_OP_NONE = 0,
@@ -172,52 +156,23 @@ struct nhg_ctx {
 };
 
 
-extern void nhg_connected_free(struct nhg_connected *dep);
-extern struct nhg_connected *nhg_connected_new(struct nhg_hash_entry *nhe);
-
-/* nhg connected tree direct access functions */
-extern void nhg_connected_tree_init(struct nhg_connected_tree_head *head);
-extern void nhg_connected_tree_free(struct nhg_connected_tree_head *head);
-extern bool
-nhg_connected_tree_is_empty(const struct nhg_connected_tree_head *head);
-extern struct nhg_connected *
-nhg_connected_tree_root(struct nhg_connected_tree_head *head);
-extern void nhg_connected_tree_del_nhe(struct nhg_connected_tree_head *head,
-				       struct nhg_hash_entry *nhe);
-extern void nhg_connected_tree_add_nhe(struct nhg_connected_tree_head *head,
-				       struct nhg_hash_entry *nhe);
-
 /**
  * NHE abstracted tree functions.
  * Use these where possible instead of the direct ones access ones.
  */
-
-
 extern struct nhg_hash_entry *zebra_nhg_resolve(struct nhg_hash_entry *nhe);
-extern uint32_t zebra_nhg_get_resolved_id(uint32_t id);
 
-/* Depends */
 extern unsigned int zebra_nhg_depends_count(const struct nhg_hash_entry *nhe);
 extern bool zebra_nhg_depends_is_empty(const struct nhg_hash_entry *nhe);
-extern void zebra_nhg_depends_del(struct nhg_hash_entry *from,
-				  struct nhg_hash_entry *depend);
-extern void zebra_nhg_depends_add(struct nhg_hash_entry *to,
-				  struct nhg_hash_entry *depend);
-extern void zebra_nhg_depends_init(struct nhg_hash_entry *nhe);
-/* Dependents */
+
 extern unsigned int
 zebra_nhg_dependents_count(const struct nhg_hash_entry *nhe);
 extern bool zebra_nhg_dependents_is_empty(const struct nhg_hash_entry *nhe);
-extern void zebra_nhg_dependents_del(struct nhg_hash_entry *from,
-				     struct nhg_hash_entry *dependent);
-extern void zebra_nhg_dependents_add(struct nhg_hash_entry *to,
-				     struct nhg_hash_entry *dependent);
-extern void zebra_nhg_dependents_init(struct nhg_hash_entry *nhe);
 
-
+/* Lookup ID, doesn't create */
 extern struct nhg_hash_entry *zebra_nhg_lookup_id(uint32_t id);
-extern int zebra_nhg_insert_id(struct nhg_hash_entry *nhe);
 
+/* Hash functions */
 extern uint32_t zebra_nhg_hash_key(const void *arg);
 extern uint32_t zebra_nhg_id_key(const void *arg);
 
@@ -243,29 +198,31 @@ extern int zebra_nhg_kernel_del(uint32_t id);
 extern struct nhg_hash_entry *
 zebra_nhg_rib_find(uint32_t id, struct nexthop_group *nhg, afi_t rt_afi);
 
-
-void zebra_nhg_free_members(struct nhg_hash_entry *nhe);
-void zebra_nhg_free(void *arg);
-void zebra_nhg_decrement_ref(struct nhg_hash_entry *nhe);
-void zebra_nhg_increment_ref(struct nhg_hash_entry *nhe);
-
-void zebra_nhg_check_valid(struct nhg_hash_entry *nhe);
-void zebra_nhg_set_if(struct nhg_hash_entry *nhe, struct interface *ifp);
-
-extern int nexthop_active_update(struct route_node *rn, struct route_entry *re);
-
+/* Reference counter functions */
+extern void zebra_nhg_decrement_ref(struct nhg_hash_entry *nhe);
+extern void zebra_nhg_increment_ref(struct nhg_hash_entry *nhe);
 extern int zebra_nhg_re_update_ref(struct route_entry *re,
 				   struct nhg_hash_entry *nhe);
 
+/* Check validity of nhe, if invalid will update dependents as well */
+extern void zebra_nhg_check_valid(struct nhg_hash_entry *nhe);
+
+/* Convert nhe depends to a grp context that can be passed around safely */
 extern uint8_t zebra_nhg_nhe2grp(struct nh_grp *grp, struct nhg_hash_entry *nhe,
 				 int size);
 
-void zebra_nhg_install_kernel(struct nhg_hash_entry *nhe);
-void zebra_nhg_uninstall_kernel(struct nhg_hash_entry *nhe);
+/* Dataplane install/uninstall */
+extern void zebra_nhg_install_kernel(struct nhg_hash_entry *nhe);
+extern void zebra_nhg_uninstall_kernel(struct nhg_hash_entry *nhe);
 
 /* Forward ref of dplane update context type */
 struct zebra_dplane_ctx;
-void zebra_nhg_dplane_result(struct zebra_dplane_ctx *ctx);
+extern void zebra_nhg_dplane_result(struct zebra_dplane_ctx *ctx);
 
-void zebra_nhg_sweep_table(struct hash *hash);
+
+/* Sweet the nhg hash tables for old entries on restart */
+extern void zebra_nhg_sweep_table(struct hash *hash);
+
+/* Nexthop resolution processing */
+extern int nexthop_active_update(struct route_node *rn, struct route_entry *re);
 #endif
